src/data_extractor/requirements.txt
openai>=1.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pymupdf>=1.24.0